Only use bitmapist-server to store statistics
Originally, stats were stored in Redis. This was extremely memory
intensive (30GB+). An alternative is bitmapist-server, which only needs
a few hundred MB to store the same data. We've now run Redis and
bitmapist-server in production in parallel for a month, with both
backends being equally consistent and stable. So, now, we are disabling
Redis (for stats - it's still used for other purposes) and are only
using bitmapist-server instead. This allows us to decrease the server's
RAM usage by 30+GB.