Auto-compute sha256 for postinstall Actions
The update client, update_engine, expects this field for postinstall
actions. To accommodate this until now, every postinstall action had to
manually have the sha256 specified via the "Other attributes" field.
This was a manually tedious and error-prone process.

This new implementation automatically computes the sha256 field on the
fly. It also migrates existing Action instances to correctly use the new
field.